Factored out splash screen to it’s own file
