+++ www/src/components/shell/CloudShell.js
@@ -3,8 +3,7 @@ import { Box } from 'grommet'
 import { useNavigate, useSearchParams } from 'react-router-dom'
 import { useMutation, useQuery } from '@apollo/client'
 import { BrowserIcon, CloudIcon, GearTrainIcon, GitHubIcon, Stepper } from 'pluralsh-design-system'
-import { Button, Div, Flex, H1, H2, P, Text } from 'honorable'
-import { CSSTransition, Transition } from 'react-transition-group'
+import { Button, Div, Flex, H1, P, Text } from 'honorable'
 
 import { LoopingLogo } from '../utils/AnimatedLogo'
 
@@ -18,6 +17,8 @@ import { Exceptions, getExceptions } from './validation'
 import { CLOUD_VALIDATIONS, ProviderForm, synopsis } from './cloud/provider'
 import { SCM_VALIDATIONS, ScmInput } from './scm/ScmInput'
 import { Github as GithubLogo, Gitlab as GitlabLogo } from './icons'
+
+import SplashToLogoTransition from './SplashToLogoTransition'
 
 const SECTIONS = {
   git: ['cloud', null],
@@ -279,8 +280,6 @@ export function OAuthCallback({ provider }) {
 
   if (!data) return <LoopingLogo dark />
 
-  console.log(data)
-
   return (
     <Box
       background="background"
@@ -417,15 +416,7 @@ export function CloudShell({ oAuthCallback }) {
   const { data: shellData } = useQuery(CLOUD_SHELL, { fetchPolicy: 'cache-and-network' })
   const [mutation] = useMutation(REBOOT_SHELL)
   const [created, setCreated] = useState(false)
-  const [splashTimerDone, setSplashTimerDone] = useState(false)
   const [stepIndex, setStepIndex] = useState(0)
-  const splashWaitTime = 1200
-
-  useEffect(() => {
-    setTimeout(() => {
-      setSplashTimerDone(true)
-    }, splashWaitTime)
-  }, [])
 
   useEffect(() => {
     if (shellData && shellData.shell && !shellData.shell.alive) {
@@ -434,99 +425,11 @@ export function CloudShell({ oAuthCallback }) {
     }
   }, [shellData, setCreated, mutation])
 
-  const showSplashScreen = useMemo(
-    () => (!shellData || !data || !splashTimerDone),
-    [shellData, data, splashTimerDone]
+  const ready = useMemo(
+    () => (shellData && data),
+    [shellData, data]
   )
   if ((shellData && shellData.shell) || created) return <Terminal />
-
-  const logoSizeBig = 48
-  const logoSizeSmall = 40
-  const logoSize = showSplashScreen ? logoSizeBig : logoSizeSmall
-
-  const fadeTransitionStyles = {
-    entering: { opacity: 1 },
-    entered: { opacity: 1 },
-    exiting: { opacity: 0 },
-    exited: { opacity: 0 },
-  }
-
-  const splashTranslate = 'calc(50vh - (110px))'
-
-  const logoTranslateTransition = {
-    '&, &.enter, &.enter-active, &.enter-done, &.exit': {
-      transform: `translateY(${splashTranslate})`,
-    },
-    '&.exit-active': {
-      transition: 'all 0.6s cubic-bezier(0.5, 0, 0.5, 1)',
-    },
-    '&.exit-active, &.exit-done': {
-      transform: 'translateY(0)',
-    },
-  }
-
-  const logoScaleTransition = {
-    '&, .enter &, .enter-active &, .enter-done &, .exit &': {
-      transform: `scale(0.${logoSizeBig})`,
-    },
-    '.exit-active &': {
-      transition: 'all 0.6s cubic-bezier(0.5, 0, 0.5, 1)',
-    },
-    '.exit-active &, .exit-done &': {
-      transform: `scale(0.${logoSizeSmall})`,
-    },
-  }
-
-  const splashEnterTransitions = {
-    opacity: 0,
-    transform: 'translateY(30px)',
-    '.appear &, .enter &': {
-      opacity: 0,
-      transform: 'translateY(30px)',
-    },
-    '.appear-active &, .enter-active &': {
-      transition: 'all 0.7s ease',
-    },
-    '.appear-active &, .enter-active &, .appear-done &, .enter-done': {
-      opacity: 1,
-      transform: 'translateY(0)',
-    },
-  }
-  const splashLogoTransitions = {
-    ...splashEnterTransitions,
-    '.appear-active &, .enter-active &': {
-      transition: 'all 0.7s ease',
-      transitionDelay: '0.1s',
-    },
-    '.appear &, .enter &': {
-      opacity: 0,
-      transform: 'translateY(30px) scale(0.5)',
-    },
-    '.appear-active &, .enter-active &, .appear-done &, .enter-done': {
-      opacity: 1,
-      transform: 'translateY(0) scale(1)',
-    },
-    '.exit &, .exit-active &, .exit-done &': {
-      opacity: 1,
-      transform: 'none',
-    },
-  }
-  const splashTextTransitions = {
-    ...splashEnterTransitions,
-    '.exit &': {
-      opacity: 1,
-      transform: 'translateY(0)',
-    },
-    '.exit-active &': {
-      transition: 'all 0.3s ease',
-    },
-    '.exit-active &, .exit-done &': {
-      opacity: 0,
-    },
-    '.exit-done &': {
-      visibility: 'hidden',
-    },
-  }
 
   return (
     <Flex
@@ -535,73 +438,12 @@ export function CloudShell({ oAuthCallback }) {
       flexDirection="column"
       mt={3}
     >
-      <CSSTransition
-        in={showSplashScreen}
-        appear
-        timeout={1000}
+      <SplashToLogoTransition
+        showSplashScreen
+        splashTimeout={1200}
+        childIsReady={ready}
       >
-        <Div
-          position="relative"
-          zIndex={1}
-          {...logoTranslateTransition}
-        >
-          <Flex
-            width="100%"
-            justify="center"
-            {...splashLogoTransitions}
-          >
-            <Div
-              width={logoSizeSmall}
-              height={logoSizeSmall}
-              transform={`scale(0.${logoSize})`}
-              {...logoScaleTransition}
-            >
-              <LoopingLogo
-                light
-                still
-                height={logoSize}
-                scale={1}
-              />
-            </Div>
-          </Flex>
-        </Div>
-          
-      </CSSTransition>
-      <CSSTransition
-        in={showSplashScreen}
-        appear
-        timeout={1000}
-      >
-        <Div
-          transform={`translateY(${splashTranslate})`}
-          position="relative"
-          width="100%"
-          height={0}
-        >
-          <H2
-            position="absolute"
-            mt={`-${logoSizeBig - logoSizeSmall}px`}
-            pt={3}
-            fontSize={60}
-            lineHeight="115%"
-            fontWeight="500"
-            letterSpacing="-1px"
-            width="100%"
-            fontFamily="'Monument Semi-Mono', 'Monument'"
-            textAlign="center"
-            {...splashTextTransitions}
-          >
-            Welcome to Plural
-          </H2>
-        </Div>
-      </CSSTransition>
-      <Transition
-        in={!showSplashScreen}
-        mountOnEnter
-        unmountOnExit
-        timeout={1000}
-      >
-        {transitionState => (
+        {ready && (
           <Div
             position="relative"
             zIndex="0"
@@ -609,10 +451,6 @@ export function CloudShell({ oAuthCallback }) {
             maxWidth={640}
             mt={2}
             px={2}
-            transition="all 0.6s ease"
-            opacity={0}
-            className={transitionState}
-            {...fadeTransitionStyles[transitionState]}
           >
             <Div mb={3}>
               <DemoStepper stepIndex={stepIndex} />
@@ -620,7 +458,7 @@ export function CloudShell({ oAuthCallback }) {
             <CreateARepoCard data={data} />
           </Div>
         )}
-      </Transition>
+      </SplashToLogoTransition>
     </Flex>
   )
 }
